Rename files with underscores in the filename
Note that because these files are templates that could be overridden, we
have to search first for the old name (with the underscore) and then the
new name when trying to locate the template. This ensures that it will
continue to work properly for current sites with the filename overrides.